requirements.txt
streamlit
openai
python-dotenv
pandas
numpy
scikit-learn
openpyxl
tiktoken